# Conflicts:
#	src/Game.Benchmarks/Benchmarks/WorldUpdateBenchmark.cs
#	src/Libraries/Game.Server/IPlayerManager.cs
#	src/Libraries/Game.Server/World/Map.cs
#	src/Tests/Game.Tests/CommandTests.cs
#	src/Tests/Game.Tests/WorldTests.cs

+++ src/Tests/Core.Tests/OutgoingPacketTests.cs
@@ -112,7 +112,7 @@ public class OutgoingPacketTests
                 .Concat(BitConverter.GetBytes((short)obj.GetSize()))
                 .Append((byte)obj.MessageType)
                 .Concat(BitConverter.GetBytes(obj.Vid))
-                .Append(obj.Empire)
+                .Append((byte)obj.Empire)
                 .Concat(Encoding.ASCII.GetBytes(obj.Message))
                 .Append((byte)0)
         );
@@ -528,7 +528,7 @@ public class OutgoingPacketTests
                 .Concat(BitConverter.GetBytes(obj.PositionX))
                 .Concat(BitConverter.GetBytes(obj.PositionY))
                 .Concat(BitConverter.GetBytes(obj.PositionZ))
-                .Append(obj.Empire)
+                .Append((byte)obj.Empire)
                 .Append(obj.SkillGroup)
         );
     }
@@ -577,7 +577,7 @@ public class OutgoingPacketTests
                 .Concat(BitConverter.GetBytes(obj.Vid))
                 .Concat(Encoding.ASCII.GetBytes(obj.Name))
                 .Concat(obj.Parts.SelectMany(BitConverter.GetBytes))
-                .Append(obj.Empire)
+                .Append((byte)obj.Empire)
                 .Concat(BitConverter.GetBytes(obj.GuildId))
                 .Concat(BitConverter.GetBytes(obj.Level))
                 .Concat(BitConverter.GetBytes(obj.RankPoints))
